Add floor to manually set tx fee
If the user manually chooses a bitcoin tx fee, it could end up below 1000 and therefor might not be relayed by the bitcoin network.

This PR was not tested yet, but is mainly intended to be the start of a discussion: should this be done or should we let the user run into this problem, if he chooses to do that? With #271 the user is at least made aware that his manually chosen fee will slightly differ from his choice.